+++ src/mariner/mixins.py
@@ -1,5 +1,5 @@
 """Utility mixins used in Mariner."""
-from typing import Dict
+from typing import Dict, Union
 
 import aiohttp
 
@@ -43,8 +43,8 @@ class ComparableMixin:  # pylint: disable=too-few-public-methods
         return self._compare(other, lambda s, o: s != o)
 
 
-class GetPageMixin:  # pylint: disable=too-few-public-methods
-    """Represent a web scraper."""
+class RequestMixin:  # pylint: disable=too-few-public-methods
+    """Mixin for HTTP requests."""
 
     user_agent = {
         "User-Agent": "Mozilla/5.0 (X11; Linux x86_64; rv:62.0) Gecko/20100101 Firefox/62.0"
@@ -53,20 +53,49 @@ class GetPageMixin:  # pylint: disable=too-few-public-methods
     async def get(
         self,
         url: str,
+        *,
         headers: Dict[str, str] = None,
         cookies: Dict[str, str] = None,
-        timeout: int = 10,
+        timeout: int = None
     ) -> str:
         """Get the requested page.
 
         Args:
             url: Url of the page to get.
+            headers: Request headers.
+            cookies: Request cookies.
+            timeout: Request timeout.
 
         Returns:
             Raw HTML page.
         """
+        return await self.request('get', url, headers=headers, cookies=cookies, timeout=timeout)
+
+    async def request(
+        self,
+        method: str,
+        url: str,
+        *,
+        data: Union[Dict, bytes] = None,
+        headers: Dict[str, str] = None,
+        cookies: Dict[str, str] = None,
+        timeout: int = 10
+    ) -> str:
+        """Make a request.
+
+        Args:
+            method: Request method.
+            url: Url of the page.
+            data: Request payload.
+            headers: Request headers.
+            cookies: Request cookies.
+            timeout: Request timeout.
+
+        Returns:
+            HTTP response.
+        """
         headers = {**headers, **self.user_agent} if headers else self.user_agent
 
         async with aiohttp.ClientSession(headers=headers, cookies=cookies) as session:
-            async with session.get(url, timeout=timeout) as response:
+            async with session.request(method, url, data=data, timeout=timeout) as response:
                 return await response.text()