+++ src/mariner/trackerplugin.py
@@ -77,6 +77,8 @@ class TrackerPlugin(mixins.RequestMixin, abc.ABC, metaclass=TrackerMeta):
         Return:
             Parsed number.
         """
+        if number == "-":
+            return 0
         squashed = number.replace(" ", "")
         return int(squashed.replace(",", ""))