+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Quote URL when opening torrent, so it cannot be interpreted as shell command.
 - Fix KickAssTorrents tracker.
+- Correctly interpret '-' as number of seeds/leeches.
 
 ## [1.3.1] - 2019-01-25